add "skip invalid trees"
So there is this huge optimization that can be made to the search algorithm. If there are trees where invalid inputs are made (which there are many of), these trees can be discarded for further seaches.
As an example imagine if the subset divides by zero. no matter what other inputs are appended to that subset, it will always be invalid. When these branches (or whatever we should call them) can be discarded, there is more room for computing potential solutions.