fix invalid tree search
there where some issues and unoptimized pieces of code that where changed. the "max_iterations" default in the UI has been changed to fit the new algorithm, which walks the input-tree more optimally and in turn uses less iterations.
I have not tested the new values for "max iterations", but I have not found a puzzle that can be solved but takes more than 10000 iterations. The new algorithm can still take several seconds to finish, but that time is reduced on average.