change brute-force search algorithm
since generating but then skipping over invalid branches when searching takes a lot of time and can be skipped, the change was to move away from generating every possible permutation of button presses.
instead, the new approach is to generate new branches from old branches that are valid, but are not a solution.
in theory, this approach should save much time in the later stages of the search, because it never gets to a point where it generates a bunch of invalid branches than need to be skipped.
the only problem is memory. since the current implementation has a queue in which the newly generated branches are stored, and by the nature of this problem there are a lot of branches to be generated, the queue is growing fast (in testing, when 100k iterations are reached, the remaining queue held over 137k branches). the constant adding and removing items alone is not optimal for a giant list, but sorting it to process shorter branches first takes time too.
because of this, both this implementation as well as the previous algorithm will have to be benchmarked to see which one is better and should be iterated upon. (maybe i'll keep both versions, but I really hope this one is just better and the old algorithm will be "deleted" (archived by git))