fix issue with buttons
Since I forgot that "lambda" and loops have an interesting interaction, I did not properly "<variable>=<variable>" proof the whole button generation code.
Because of this, some buttons could not be added to the "buttons_availible" list but instead added uses to the last button generated in the same loop.