add break function for brute-force loop
The brute force loop can take a long time to finish, for example when I set it to 1M iterations because it did not find a solution at 100k.
When the algorithm finds any solutions but you don't want to wait for the rest of the iterations, there would be the need for a "return solutions now" function.
This is the new addition. The "KeyboardInterrupt" exception is now used as the trigger for this break function.