remove buttons that don't exist in the game
As I learned, the current version of the game does not have most of the "replace" buttons (A->B). There are only 10, which are placed in a sort of loop.
The "replace" buttons take up a lot of space in the GUI and it was diffictult to find the buttons that you actually got, so I decided to remove the buttons that do not exist in the game.