revert cost values
This is because the program will avoid using buttons that lead to usable solutions, meaning it will try all the button combinations that are the least likely to solve the game-state.
A possible solution to this dilemma would be to "reverse" the brute-forced cost for the search but use it for displaying the solutions.
This solution may be implemented by someone in the future, but this commit is just a "hotfix" (the problem existed for quite a while now, so the problem is rather "cold" by now) for the problem.